Added session records for new tests
