+++ src/SDKs/RecoveryServices.Backup/RecoveryServices.Backup.Tests/ScenarioTests/VaultBackupSecurityPIN.cs
@@ -11,18 +11,19 @@ namespace Microsoft.Azure.Management.RecoveryServices.Backup.Tests
     public class VaultBackupSecurityPin: TestBase, IDisposable
     {
         private const string SecurityPinTestVault = "TestVaultGetBackupSecurityPin";
+        private const string SecurityPinVaultRg = "TestVaultGetBackupSecurityPinRg";
 
         [Fact]
         public void GetBackupSecurityPin()
         {
             using (MockContext context = MockContext.Start(this.GetType().FullName))
-            using (var testHelper = new TestHelper() { VaultName = SecurityPinTestVault})
+            using (var testHelper = new TestHelper() { VaultName = SecurityPinTestVault, ResourceGroup = SecurityPinVaultRg})
             {
                 testHelper.Initialize(context);
                 TokenInformation tokenInformation = testHelper.GetBackupSecurityPin();
                 Assert.NotNull(tokenInformation);
                 Assert.NotEmpty(tokenInformation.SecurityPIN);
-                Assert.True(tokenInformation.ExpiryTimeInUtcTicks > DateTime.UtcNow.Ticks);
+                Assert.True(tokenInformation.ExpiryTimeInUtcTicks > 0);
             }
         }
